fix progress bugs

+++ HCMVManager/MediaEdit/MediaAction.h
@@ -23,7 +23,7 @@ typedef NS_ENUM(int, SType)
 
 /* Action 模板，即Action的抽像操作模式 */
 @interface MediaAction : HCEntity
-@property (nonatomic,assign) int    MediaActionID;
+@property (nonatomic,assign) long    MediaActionID;
 @property (nonatomic,PP_STRONG) NSString * ActionTitle;
 @property (nonatomic,PP_STRONG) NSString * ActionIcon;
 @property (nonatomic,assign) SType ActionType;//暂定4个，1 表示慢速 2 表示加速 3表示Rap 4表示倒放 0表示无特效 99999表示是一个模板类型的，
